pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "urbanpulse"
version = "0.1.0"
description = "Real-Time Socioeconomic Sensing via Multi-Modal Urban Data"
authors = [{name = "Vidhi Kadam"}]
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"

dependencies = [
    "numpy>=1.24",
    "pandas>=2.0",
    "scikit-learn>=1.3",
    "scipy>=1.11",
    "torch>=2.1",
    "torchvision>=0.16",
    "xgboost>=2.0",
    "geopandas>=0.14",
    "osmnx>=1.7",
    "pyyaml>=6.0",
    "requests>=2.31",
    "tqdm>=4.66",
    "matplotlib>=3.8",
    "pyarrow>=14.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "black>=23.12",
    "isort>=5.13",
    "flake8>=6.1"
]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
